capital_quiz/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
python-decouple==3.8
typing-extensions==4.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
django-decouple==2.1
django-filter==23.2
django-postgres-extra==2.0.8
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.8.0
libretranslatepy==2.1.1
lxml==4.9.3
Markdown==3.4.4
pkg_resources==0.0.0
psycopg==3.1.10
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
translate==3.6.1
typing_extensions==4.7.1
urllib3==2.0.4
zipp==3.16.2
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.4
django-decouple==2.1
django-filter==23.2
django-postgres-extra==2.0.8
djangorestframework==3.14.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==3.4
importlib-metadata==6.8.0
libretranslatepy==2.1.1
lxml==4.9.3
Markdown==3.4.4
pkg_resources==0.0.0
psycopg==3.1.10
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
translate==3.6.1
typing_extensions==4.7.1
urllib3==2.0.4
zipp==3.16.2
